+++ app/Http/Controllers/OpenBanking/CoinbaseController.php
@@ -4,12 +4,11 @@ namespace App\Http\Controllers\OpenBanking;
 
 use App\Enums\BankingProvider;
 use App\Http\Requests\OpenBanking\ConnectCoinbaseRequest;
-use App\Models\User;
 use App\Services\AccountUserCurrencyService;
 use App\Services\Banking\CoinbaseClient;
 use Illuminate\Http\JsonResponse;
 
-class CoinbaseController extends OpenBankingConnectController
+class CoinbaseController extends CryptoPortfolioConnectController
 {
     /**
      * Validate Coinbase CDP API credentials and create a connection.
@@ -34,11 +33,6 @@ class CoinbaseController extends OpenBankingConnectController
         return 'https://whisper.money/storage/banks/logos/coinbase.png';
     }
 
-    protected function aspspCountry(array $validated): string
-    {
-        return $validated['country'];
-    }
-
     protected function fetchProviderData(array $validated): mixed
     {
         $client = new CoinbaseClient($validated['api_key_name'], $validated['private_key']);
@@ -55,16 +49,5 @@ class CoinbaseController extends OpenBankingConnectController
     protected function credentialErrorMessage(\Throwable $e): string
     {
         return 'Invalid API credentials or failed to connect to Coinbase.';
-    }
-
-    protected function buildPendingAccounts(mixed $providerData, User $user): array
-    {
-        return [
-            [
-                'uid' => 'coinbase-portfolio',
-                'currency' => $user->currency_code,
-                'name' => 'Crypto Portfolio',
-            ],
-        ];
     }
 }
